pyproject.toml
[project]
name = "lang-graph"
version = "0.1.0"
description = "LangGraph sample project"
requires-python = ">=3.12"
readme = "README.md"
dependencies = [
    "langgraph>=0.2.0",
    "langgraph-cli[inmem]>=0.1.0",
    "langgraph-runtime-inmem==0.22.1",
    "langchain>=0.3.0",
    "langchain-huggingface>=0.1.0",
    "transformers>=4.47.0",
    "torch>=2.0.0",
    "accelerate>=1.2.0",
    "sentencepiece>=0.2.0",
    "python-dotenv>=1.0.0",
    "langsmith>=0.2.0",
    "tqdm>=4.66.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
package = true
